[IOT-2539] Removed unused-param warning
Now that -Werror is turned on for JNI code the error
must be fixed to build. This was missed earlyer
because it is only build when WITH_CLOUD=1 is in the
scons options which is not a default build option.

Bug: https://jira.iotivity.org/browse/IOT-2539
Change-Id: I7790cba1baa140c3059f0a91b66882425ed5a6cb
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/21645
Reviewed-by: Larry Sachs <larry.j.sachs@intel.com>
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Rick Bell <richard.s.bell@intel.com>
(cherry picked from commit db7b6d26c941525c0e8428f70864905997861c90)
Reviewed-on: https://gerrit.iotivity.org/gerrit/22215