[IOT-2539] Clean sign-compare warnings
It is common to do a compair with a const value in unit
tests. For example EXPECT_EQ(3, value).  If `value` is an
unsigned type this will produce the sign-compare compiler
warning. The compiler defaults to treating constant
numbers as a signed data type.  Adding a sufix u (i.e. 3u)
will tell the compiler to treat the constant as an
unsigned data type.

In other cases the type is cast to an unsigned or signed
value before the comparison.

In aclresourcetest.cpp the return type for GetNumberOfResource
function was updated to return size_t since in all the test
conditions it is compared with variable of size_t or a
constant number.

Bug: https://jira.iotivity.org/browse/IOT-2539
Change-Id: I7e47bfb9f027304abd975eb0d8e577cbd484545a
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/21689
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>