[IOT-2539] Clean implicit-function warning
pmutility.c code would give an implicit-funtion-declaration
warning when the usleep function was used. On linux the
usleep function is part of unistd.h which was already
included.

usleep is not exposed unless _DEFAULT_SOURCE or _BSD_SOURCE
is defined before including unistd.h. To prevent another warning
[-Wcpp] both macros are included. As directed in the glibc
release notes:
https://sourceware.org/glibc/wiki/Release/2.20#Packaging_Changes

Bug: https://jira.iotivity.org/browse/IOT-2539
Change-Id: I7002ef303ba685bda76f70590d7c9593435b4ba7
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/21707
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>