IOT-2539 remove unused function warning
The static function GenerateEccKeyPair in occertutility.c
in unused and can be delete.

Note: there is a line for line copy of this function also
found in certhelpers.c that is used. So deleting this
function will not cause the code to be lost.

Bug: https://jira.iotivity.org/browse/IOT-2539
Change-Id: I3a5aefdd316cf5cedd6af58811358ebdd3589838
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/21735
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Phil Coval <philippe.coval@osg.samsung.com>
Reviewed-by: Rick Bell <richard.s.bell@intel.com>
(cherry picked from commit b678dcbb8be1b0eb00ba48625ba511fa2412901c)