IOT-2539 removed unused code from subownerclient.c
The subowner client contained several functions and
variables that were unused. Unused code is not
maintained and should be removed.

selectTwoDiffNum is an unused function
printUuidList is an unused function
updateDoxmForMOTCB is an unused function
g_unown_cnt is an unused variable
SUPPORTED_PRMS is an unused const varaible

provisionCred is unused but unlike the other unused
code it was referenced in some commented out code
in main.  For that reason the provisionCred code
was not just removed. Instead a macro was used to
block out the code from the compiler. The code can
be enabled by changing PROVISION_CRED_SUPPORTED macro
to 1.

Change-Id: Ia89d1ebd4cfeff7e8771147d309ba7155f97818c
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/21755
Reviewed-by: Alex Kelley <alexke@microsoft.com>
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>
Reviewed-by: Randeep Singh <randeep.s@samsung.com>