[IOT-2598][EasySetup]Handle oic.if.rw,oic.if.r
Add handling for "oic.if.rw" for wificonf, coapcloudconf.
Add handling for "oic.if.r" for devconf.
Slightly refactor ProcessGetRequest to move out common part.
Update CheckEhRequestPayload to check for all supported interfaces in
generic manner.
Update "if" property of response of resources. (This is still work in
progress for response of EasySetup Resource, will be updated in a
separate commit).

Change-Id: I501a6ba3aff8e84c253f962d383782b2c6f49206
Signed-off-by: Ankur Bansal <ankur.b1@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/22265
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Abhishek Pandey <abhi.siso@samsung.com>
Reviewed-by: Ashok Babu Channa <ashok.channa@samsung.com>