security: Advertise certificate creds support
Allow the OBT to discover that the current device/app supports certificate
based credentials.

Note that this is an improvement just for those devices/apps generating a
default *.dat file when they start for the first time. Similar work will be
required for the IoTivity sample apps that use custom-made *.dat files.

CT1.7.4.2 is passing after this change.

Change-Id: Ia63d63e74bf637c376055b5c14ac4620304eb799
Signed-off-by: Dan Mihai <Daniel.Mihai@microsoft.com>
bug: https://jira.iotivity.org/browse/IOT-2604
Reviewed-on: https://gerrit.iotivity.org/gerrit/21865
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Nathan Heldt-Sheller <nathan.heldt-sheller@intel.com>