Improve on test/valgrind setup
Problem: when Jenkins runs the unit_test configuration on Linux,
it looks for files suffixed .xml in the top directory, but it
does not know where they come from, and in a multi-stage run (as
unit_test currently is), since the collection happens only at the
end, there can be some confusion about which results apply to the
"failure count".

As a first stage in cleanup, to avoid programmer errors when
calling run_tests, the filename for the valgrind run is now
generated from the pathname of the test.  Until further cleanup,
the passed filename is ignored and just used as a boolean (that
aspect was already documented - empty string meaning do not
use valgrind).

Whether or not valgrind runs happen was selected based on platform.
An environment flag is added to allow it to be turned of for
small-memory platforms, or for testing whether valgrind is adding
problems (that is, if unit tests fail, can try again with valgrind
disabled to help narrow down where the problem is).

As a note, to combat the "overwrite" effect when unit_test
builder runs non-secure followed by secure build+test, was
attempting to separate those two builds by putting them in
distinct paths, but apparently the path is computed other places
than build_common/SConscript.  android, darwin and tizen all found
some existing instance of the current path style even after it
was changed and so broke.  So as of this commit, this topic is
not addressed.

Change-Id: I57213dce98a9a0f8ced3a18e82d0a51691b024bd
Signed-off-by: Mats Wichmann <mats@linux.com>