generic-java: remove unbuilt samples
All of the Android samples were copied over to the
generic-java samples folder but only 3 of the samples
were actually updated so they would build and run.
Till there is time to actually work through the
samples and make sure they build and run they
should be removed so they are not referenced by
developers.

Change-Id: I64ef0f2d5d10e3aa66f85f4f34b9e0cd29d94542
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/22377
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Larry Sachs <larry.j.sachs@intel.com>
Reviewed-by: Rick Bell <richard.s.bell@intel.com>