Add CONTRIBUTING.md
github evaluates the maturity of a project in part by whether it has
certain things set up: a readme, a license, a contributions document
and a code of conduct.  There's no cost to adding the CONTRIBUTING.md
which the github mirror will pick up - it's the content from a web page
(and points to that page as well).

Change-Id: Idb77d3b822482ddc430cf500eca0a6bae5ccc955
Signed-off-by: Mats Wichmann <mats@linux.com>