IOT-2678 Change Java back to dynamic linking
Java was changed to use static linking for most of
the libraries. This was done to enable building Java
on Windows which does not contain the same dynamic
libraries as are avalible to Linux and Android.

See IOT-2405

Unfortanatly this change caused an issue in the Android
serivces. It caused multiple instances of the IoTivity
stack to be included in different libraries used by
the Android services.

This reverts the key elements of IOT-2405 without
reverting the entier change.

Relates-to: https://gerrit.iotivity.org/gerrit/21061

This is being done till we can figure out how to
enable the build on windows so it does not cause
multiple copies of the oc stack.

Windows java will no longer build if this commit
is accepted. However, Android is considered a
higher priority OS when it comes to the Java
language.

Bug: https://jira.iotivity.org/browse/IOT-2678
Bug: https://jira.iotivity.org/browse/IOT-2405
Change-Id: I54e9e65b00a199151395542d9e4629cef4efe516
Signed-off-by: George Nash <george.nash@intel.com>