Fix build problem with connectivity_abstraction
A linux system started failing builds making the provisioning unit test.
This file links with libconnectivity_abstraction_internal.  However,
other libraries in the link line require libconnectivity_abstraction, so
the linker decides it's an error for it not to be listed on the link line.

Fix: flip the library requested to the public libconnectivity_abstraction,
except for Windows where only the internal one is available.
It's done this way in two other build scripts; it's possible it
should be done in the others that don't although this is the
only one causing a build fail in this instance.  (candidates:
resource/csdk/security/provisioning/unittest/SConscript.
resource/csdk/security/unittest/SConscript,
resource/csdk/stack/test/SConscript, resource/unittests/SConscript)

Change-Id: I7b86d9eafa708e0b59a504d1c240d641134c5e48
Signed-off-by: Mats Wichmann <mats@linux.com>