build: make linux shared libs with as-needed flag
The shared libraries in the Linux build sometimes link with other shared
libraries they don't actually use.  It's "cleaner" if the dependencies
are only real ones.  Adding the linker flag only for shared library builds
(SHLINKFLAGS) improves this situation, and leads to the changes documented
in the referenced bug (IOT-2354).

The scons scripts which build these libraries should presumably
be examined, to stop them listing dependencies they don't use,
but that is currently a separate topic.

Change-Id: I0b55c2048af36a28ae26829aa934cfadc4665651
Bug: IOT-2354
Signed-off-by: Mats Wichmann <mats@linux.com>