build: variant-dir fix in IPCA
One more place turned up where build happens in src dir instead of
in variant dir, which impacts parallel builds of different flavors.
It seems that placing a source file from another part of the
project into the source list defeats the variant dir.  This proposal
takes it out and linked to the library it appears in instead.

Change-Id: I73283fb854fd899e5609017f2bb8a52fc8c00675
Signed-off-by: Mats Wichmann <mats@linux.com>