IOT-2539 Enable the -Werror build flag
The coding standard states:
Code must compile with no warnings. Without flags
forcing warnings off.
see:
https://wiki.iotivity.org/iotivity_c_coding_standards

Its possible that using an untested version of gcc could
result in build warings that are not yet fixed. For this
reason the build option ERROR_ON_WARN was added to make
it possible to turn off the -Werror flag. This should
only be used locally an not changed on the CI build.

Due to some bugs in older version of gcc some warnings
are forced off when using an older version of gcc.

This change will force code to follow the coding
standard of building with no warnings.

The coapHttpParser code produces a build warning
that currently is not solved. When building the
coap_http_proxy library the -Werror build flag is
removed.

Don't use the -Werror build option for extlib yaml
code.

Don't use the -Werror build option for coap library

Change-Id: Ifcc25ed7e5b8637ac4383a7bfa51ace105ed9458
Signed-off-by: George Nash <george.nash@intel.com>