IOT-2539 Targeted ignore of type-limits warning
This will turn of the type-limits warning in very
targeted area in the code. The code is checking
that the user passed in value is within limits.

Although it is very unlikely to be outsite the
expected values it is possible. Tthis selectivly
shuts off the warning for the two functions the
warning is found in.

Bug: https://jira.iotivity.org/browse/IOT-2539
Change-Id: Id5da64534c05dffd479d911112d5569305c36d48
Signed-off-by: George Nash <george.nash@intel.com>