[IOT-2710][IOT-2806] fix /pstat Update handler
The way CBOR->/pstat binary is written, the Properties in the
current gPstat are added to the binary representation even if
the CBOR payload doesn't contain that Property. This may be ok
in most cases since it means that (if a Property is omitted)
the gPstat value won't be changed from whatever it is.
But this also means that if /pstat.dos.s is Updated,
the /pstat.dos.s Property Update must be done after all other
Properties, so that if a /pstat.dos.s update changes other
/pstat Properties, the new values aren't subsequently overwritten.

Change-Id: I07369355ec3141c40b5178cae41ba02e5e7813fc
Signed-off-by: Nathan Heldt-Sheller <nathan.heldt-sheller@intel.com>