IOT-2539 Fix -Wswitch simulator
Fixed the switch warning in the simulator code.

In the request_model_builder the other posible values
for the switch are not expected if they occure print
an error and exit assert. Just in case it is a release
build an empty string is returned.

In simulatory_utils the un-handled case statments were
added to the switch.

Bug: https://jira.iotivity.org/browse/IOT-2539
Change-Id: Iebb79bd22410727c65c33ed73ec8704a1b8a33a4
Signed-off-by: George Nash <george.nash@intel.com>