[IOT-2820] make /pstat.dos.p read-only
The Parameter "p" in the /pstat "dos" Property is meant
to be read-only in every Device Onboarding State.  This
fixes the Update handler to treat "p" as read-only.  It
also fixes the pstat->CBOR marshaling code to provide the
option to omit "p". And it updates the provisioningclient
code to omit "p" from Updates to /pstat.

Change-Id: I25a92bd33e5476d7d5896021e42102762deb6c31
Signed-off-by: Nathan Heldt-Sheller <nathan.heldt-sheller@intel.com>