IOT-2539 clean unused-function in caleadapter.c
The static functions:
  CAInitLEServerQueue
  CAInitLEServerReceiverQueue
  CAInitLEServerSenderQueue
  CALEServerDataReceiverHandler
  CALEServerSendDataThread

Are all only called if both ROUTING_GATWAY and
SINGLE_THREAD are not defined. When the are not
called gcc will produce the -Werror=unused-function
warning.

Rather than spatering the macros all over the file
I moved all of the functions into a single block

The code in the functions are not modified.

Bug: https://jira.iotivity.org/browse/IOT-2539
Change-Id: Ic15e4e91028b87fb8d2680f2e64c724d1fae25cc
Signed-off-by: George Nash <george.nash@intel.com>