[IOT-2821] Fix /doxm Update handling
The Update handler is still doing things such as
self-updating /doxm.rowneruuid or /doxm.owned.  Also, the
logic for determining OTM start/stop is wrong.  This patch
cleans up the JustWorks flow as a template, and the same
fix will be applied to Random PIN and Cert-based once
verified and reviewed.

Change-Id: Ieb68911de87cf816597352aec760b3ef85baa87d
Signed-off-by: Nathan Heldt-Sheller <nathan.heldt-sheller@intel.com>