[IOT-2837] Enable IoTivity build on Raspbery pi
The default arch for the raspberry pi board is either armv6l or
armv7l depending on board hardware version. Neither of them are
included in the valid values for the TARGET_ARCH build option.
The current workaround is to pass TARGET_ARCH=arm to the build
command. This change will enable the default arch.

Change-Id: I7f07734f5e06d79d13275254fb01f924227e229a
Signed-off-by: Rami Alshafi <ralshafi@vprime.com>