IOT-2838 Fix javadoc warnings
Documented all the functions referencing the native code.

Note the OcProvisioning setPinType method had a return value
int. The was returning the return value of the native call
that was the status. Due to the way the code was writen the
only value that could be returned was when things worked as
expected. All other values were thrown so the return value
was un-needed. The function was changed to have a return type
of void.

Bug:https://jira.iotivity.org/browse/IOT-2838
Change-Id: I3b5955a336661574dde9f70c87b6ead3dcd5ea7b
Signed-off-by: George Nash <george.nash@intel.com>