Adapt gcc version-compare to deal with gcc7, py3
For gcc starting with 7, gcc -dumpversion returns a single (non
dotted) number. This is actually an illegal version to distutils'
version.StrictVersion, so use LooseVersion instead, where that is a
legal number but the comparison still works fine.

For Python 3.x, subprocess.check_output returns a byte string,
which needs to be decoded before it can be passed to the version
comparator function. Instead of bothering with the decoder, which is
hard to get right so it works for both Py2 and Py3 case, call the new
subprocess.getoutput() method, falling back to check_output if not
available.

Change-Id: I29e1f43f8fbde130103099be1972b574a23442cf
Bug: https://jira.iotivity.org/browse/IOT-2862
Signed-off-by: Mats Wichmann <mats@linux.com>