IOT-2848 Fix for intermitent read error
Valgrind was reporting intermitent read errors
from the provisiontest. I think the result of the
failure was a change made for IOT-2599 the name
of the tests db file was changed from ./dbpath
to ./dbpath.tmp however the name of the file was
used in more than one test.
See: https://gerrit.iotivity.org/gerrit/#/c/22057/

Bug: https://jira.iotivity.org/browse/IOT-2848
Bug: https://jira.iotivity.org/browse/IOT-2599
Change-Id: Id72a07b51aee10abe34a428239ca74dfbf7a2641
Signed-off-by: George Nash <george.nash@intel.com>