IOT-2826 Update *.dat files for samples
Both client and the server used the incorrect encoding
for the certificate data. When the server .dat file was
updated with the fix from
https://gerrit.iotivity.org/gerrit/22535/ I did not
update the client .dat file because I had not made any modifications to the client. This then caused the
certificates to miss match. If I run json2cbor on both the client and the server *.json input files the error no
longer occurs.

Bug: https://jira.iotivity.org/browse/IOT-2826
Change-Id: I8d5acc8a49533b7cca7b5c4f46371e495449aa73
Signed-off-by: George Nash <george.nash@intel.com>