Windows: Enable UWP to be built on master
If you are building for UWP, then you can't build and run a regular
desktop program/exe. Therefore, needed tools like json2cbor will
fail to run and the build will fail to generate the .dat files that
are needed.
- Change run.bat to build the win32 version of the security tool
(resource/csdk/security/tool) which will be copied to the uwp output
directory to use for the build.
- Change the security tool SConscript file to check for this when
building for UWP.

Change-Id: I2d83cdf3d78cc73a9e31be2741251d533c1ae93a
Signed-off-by: Ibrahim Esmat <iesmat@microsoft.com>